+++ FE/src/components/Common/index.ts
@@ -1,5 +1,4 @@
 export { default as Button } from "./Button/Button";
-export { default as Image } from "./Image/Image";
 export { default as Spacing } from "./Spacing/Spacing";
 export { default as FavoriteBtn } from "./FavoriteBtn/FavoriteBtn";
 export { default as Overlay } from "./Overlay/Overlay";
