+++ FE/src/components/ui/index.ts
@@ -1,3 +1,4 @@
 export { default as Text } from "./Text/Text";
 export { default as Heading } from "./Heading/Heading";
 export { default as Input } from "./Input/Input";
+export { default as Image } from "./Image/Image";
